+++ flowman-core/src/main/scala/com/dimajix/flowman/jdbc/SqlCommands.scala
@@ -30,6 +30,8 @@ abstract class SqlCommands {
 
     def getJdbcSchema(statement:Statement, table:TableIdentifier) : Seq[JdbcField]
 
+    def updateComment(statement:Statement, table: TableIdentifier, column:String, comment:Option[String]) : Unit
+
     def getStorageFormat(statement:Statement, table:TableIdentifier) : Option[String]
     def changeStorageFormat(statement:Statement, table:TableIdentifier, storageFormat:String) : Unit